Command line interface improved with argparse library.
More options are now available. Still need some work on the plot commands and input formats.
Multi-component heatcurrent: bug fixed - the code was not able to identify a (1,X,Y) array as a 1-component fluid with Y cartesian components.
Cepstral and Resample log messages are now stored in the containing objects.