New `MDSample.resample` method, substitutes `resample_currents`
The new `MDSample.resample` method performs the filter & resample that
was previously performed by `resample_currents`. It calls the function
`resample_timeseries` defined in `md/resample.py`.
This works for any MDSample-derivated class, e.g. a `HeatCurrent`.
It uses the `_get_builder` method of the class to get the parameters
needed to define a new object of the same type, with `trajectory` and
`DT_FS` updated.

To resample a time series, you just need to call its method:
  j_new = j.resample(TSKIP, ...)

The trajectory read by `resample_timeseries` can be 1D, 2D, or 3D. In
the latter case (multi-current), the `filter_and_sample` is applied to
each row in the first dimension (each current) separately, and the 3D
trajectory is reconstructed afterwards.

TODO: the plot functions of `resample_timeseries` are disabled, because
`MDSample` does not have a `plot_periodogram` method. For now plots are
performed in the `HeatCurrent.resample` method.