`Current` base class introduced, new obj definition, units moved
The `Current` becomes the base class for any current associated to a
transport coefficient. `HeatCurrent` is now subclassed from it.
These classes have been separated and moved to `current/` directory.

A `Current` is defined from a trajectory and a undefined number of
key arguments (**params). The precise list of parameters (constants)
needed by a current can be specified by defining the `_input_parameters`
and `_optional_parameters` class attributes.

NOTE: it is not possible to define a Current object with positional
arguments. Apart from the trajectory, all the other parameters must be
passed as key arguments. E.g.
  j = HeatCurrent(traj, DT_FS=1.0, TEMPERATURE=300., VOLUME=100.,
                  UNITS='metal')
These key arguments are case-insensitive.

To define a new current, one just needs to define the `_get_builder`,
`get_units_list`, and `initialize_units` methods, at least, which are
virtual (not defined in the `Current` class).

TODO: find a way to scan the folder `current/units` to get a list of
available units for a particular `_current_type` (heat, electric, ...).