Fix bug with units and `KAPPA_SCALE`
Units included `DT_FS`, which should not be included, because it is
already in the definition of the psd.
Tests, plotter, analysis were using the lowercase `kappa_scale`.

TODO: test_as_example does NOT work because results are compared with a
numerical precision that is too high. Different machines will give slightly
different numerical results!