New Plotter system and CurrentPlotter implemented
The plotter module has been reorganized (#25).
Now it contains all the plot functions. A `Plotter` subclass can be
defined by simply importing the relevant functions.
`CurrentPlotter` has been defined: it contains all the plot functions
used by `Current`.

The `Current` class has a `_default_plotter` attribute (`CurrentPlotter`).
The plot methods are defined dynamically by the `set_plotter` method.
This method takes all the 'plot_*' functions contained in the `plotter`
input parameter (or the default plotter, e.g. `CurrentPlotter`) and
converts them into methods of `Current`. This is obtained thanks to a
decorator (`add_method`) that was defined. The first argument of the
plot functions (`current`) becomes `self` when converted into a class
method.

The plotter of `Current` is initialized by default to `CurrentPlotter`
when importing the library. It is possible to change it dynamically by
calling the `set_plotter` method. This is a class method, hence it will
change the plotter assigned to the `Current` class and affect all the
current object already defined.
(Notice that calling `set_plotter` from a subclass will modify also the
plotter of the `Current` base class. If this behavior might be
confusing, if it is not wanted in some applications, we should change it.)

It seems that #40 is not the case anymore.

TODO:
- check and remove the commented plot methods from `mdsample` (#43).
- fix the analysis CLI. It is broken now.

In order to close #25 :
- should we define other plotters for the CLI and GUI?
  In case, evaluate shared functionalities.
- how to define plots customizations?