Simplify Current builder, by defining _builder property
Now any sub-class of `Current` needs only to define a `_builder`
property, that returns a dictionary containing the keyworded input
parameters necessary to initialize a new and identical current object.
The trajectory does not need to be specified, as it is added to this
dictionary by `_get_builder()`. This implementation removes the
copy-pasted lines defining the trajectory array, which is the same
for any kind of current.