Add LAMMPS Silica input/output files
Fixes #18
Contributes to #23
The LAMMPS input files used to generate the silica example data have
been added to the `examples/data/Silica/lammps` folder.
In the same folder, the output of the calculation is reported, along
with a script to convert it to `table` and `dict` formats.
The data had to be regenerated, therefore all the data files in the
example folder have been updated.